requirements.txt
streamlit
groq
google-genai
pandas
Pillow
python-dotenv